ref(ost): write_backlog_capacity is size_t, not uint64_t
It bounds Session::_pending_writes, real buffers held in process
memory -- unlike RawstorObjectSpec::size (object size on the backing
store, decoupled from the host's own address space), a backlog cap
can never usefully exceed what size_t can already express: nothing
could allocate more than that anyway. rawstd_size_to_bytes() itself
stays uint64_t (it also serves cli/'s object-size parsing, which does
need the wider range) -- main.cpp now parses into a local uint64_t and
narrows once parsing succeeds.

+++ ost/include/ost/server.hpp
@@ -5,7 +5,7 @@
 
 #include <rawstor/rawio.h>
 
-#include <cstdint>
+#include <cstddef>
 #include <memory>
 #include <string>
 #include <unordered_map>
@@ -23,7 +23,7 @@ private:
     std::vector<rawstd::URI> _locations;
     RawIOEvent* _accept_event;
     unsigned int _write_throttle_limit;
-    uint64_t _write_backlog_capacity;
+    size_t _write_backlog_capacity;
     std::unordered_map<int, std::shared_ptr<Session>> _sessions;
 
     static int _accept(int result, void* data) noexcept;
@@ -33,7 +33,7 @@ private:
 public:
     Server(
         unsigned int queue_size, unsigned int write_throttle_limit,
-        uint64_t write_backlog_capacity, const std::string& addr,
+        size_t write_backlog_capacity, const std::string& addr,
         unsigned int port, const char* location
     );
     Server(const Server&) = delete;
@@ -55,7 +55,7 @@ public:
 
     // Per-session cap, in bytes, on writes queued behind
     // write_throttle_limit() but not yet dispatched -- see Session::_write().
-    inline uint64_t write_backlog_capacity() const noexcept {
+    inline size_t write_backlog_capacity() const noexcept {
         return _write_backlog_capacity;
     }
