+++ ost/include/ost/session.hpp
@@ -54,7 +54,7 @@ private:
     // _write()'s use of it against Server::write_backlog_capacity() to reject
     // a write outright rather than let it grow _pending_writes without
     // bound.
-    uint64_t _pending_writes_bytes;
+    size_t _pending_writes_bytes;
 
     // Arms the multishot recv; only ever called once, from the
     // constructor.
@@ -140,7 +140,7 @@ public:
     }
 
     // For ost/tests/ to verify the write backlog cap (see _write()).
-    inline uint64_t pending_writes_bytes() const noexcept {
+    inline size_t pending_writes_bytes() const noexcept {
         return _pending_writes_bytes;
     }
 };
